+++ network/network_test.go
@@ -17,22 +17,22 @@ var _ = Describe("Network", func() {
 			activation.StepFunction,
 		).WithLayer(
 			[]network.Neuron{{
-				Weights:   []int{0, 1},
-				Threshold: 1,
+				Weights:   []float64{0.0, 1.0},
+				Threshold: 1.0,
 			}, {
-				Weights:   []int{1, 1},
-				Threshold: 2,
+				Weights:   []float64{1.0, 1.0},
+				Threshold: 2.0,
 			}, {
-				Weights:   []int{1, 0},
-				Threshold: 0,
+				Weights:   []float64{1.0, 0.0},
+				Threshold: 0.0,
 			}},
 		).WithLayer(
 			[]network.Neuron{{
-				Weights:   []int{0, 1, 0},
-				Threshold: 1,
+				Weights:   []float64{0.0, 1.0, 0.0},
+				Threshold: 1.0,
 			}, {
-				Weights:   []int{1, 0, 1},
-				Threshold: 2,
+				Weights:   []float64{1.0, 0.0, 1.0},
+				Threshold: 2.0,
 			}},
 		).Build()
 
@@ -47,13 +47,13 @@ var _ = Describe("Network", func() {
 			activation.StepFunction,
 		).WithLayer(
 			[]network.Neuron{{
-				Weights:   []int{0, 1},
-				Threshold: 1,
+				Weights:   []float64{0.0, 1.0},
+				Threshold: 1.0,
 			}},
 		).Build()
 
 		Expect(err).To(BeNil())
-		_, err = n.Run([]int{1})
+		_, err = n.Run([]float64{1.0})
 		Expect(err).To(Equal(errors.New("input values does not match input neuron count")))
 	})
 
@@ -64,46 +64,46 @@ var _ = Describe("Network", func() {
 			activation.StepFunction,
 		).WithLayer(
 			[]network.Neuron{{
-				Weights:   []int{0, 1},
-				Threshold: 1,
+				Weights:   []float64{0.0, 1.0},
+				Threshold: 1.0,
 			}, {
-				Weights:   []int{1, 1},
-				Threshold: 2,
+				Weights:   []float64{1.0, 1.0},
+				Threshold: 2.0,
 			}, {
-				Weights:   []int{1, 0},
-				Threshold: 0,
+				Weights:   []float64{1.0, 0.0},
+				Threshold: 0.0,
 			}},
 		).WithLayer(
 			[]network.Neuron{{
-				Weights:   []int{0, 1, 0},
-				Threshold: 1,
+				Weights:   []float64{0.0, 1.0, 0.0},
+				Threshold: 1.0,
 			}, {
-				Weights:   []int{1, 0, 1},
-				Threshold: 2,
+				Weights:   []float64{1.0, 0.0, 1.0},
+				Threshold: 2.0,
 			}},
 		).Build()
 
 		Expect(err).NotTo(HaveOccurred())
 
-		r, err := n.Run([]int{0, 0})
+		r, err := n.Run([]float64{0.0, 0.0})
 		Expect(err).To(BeNil())
-		Expect(r[0]).To(Equal(0))
-		Expect(r[1]).To(Equal(0))
+		Expect(r[0]).To(Equal(0.0))
+		Expect(r[1]).To(Equal(0.0))
 
-		r, err = n.Run([]int{1, 0})
+		r, err = n.Run([]float64{1.0, 0.0})
 		Expect(err).To(BeNil())
-		Expect(r[0]).To(Equal(0))
-		Expect(r[1]).To(Equal(0))
+		Expect(r[0]).To(Equal(0.0))
+		Expect(r[1]).To(Equal(0.0))
 
-		r, err = n.Run([]int{0, 1})
+		r, err = n.Run([]float64{0.0, 1.0})
 		Expect(err).To(BeNil())
-		Expect(r[0]).To(Equal(0))
-		Expect(r[1]).To(Equal(1))
+		Expect(r[0]).To(Equal(0.0))
+		Expect(r[1]).To(Equal(1.0))
 
-		r, err = n.Run([]int{1, 1})
+		r, err = n.Run([]float64{1.0, 1.0})
 		Expect(err).To(BeNil())
-		Expect(r[0]).To(Equal(1))
-		Expect(r[1]).To(Equal(1))
+		Expect(r[0]).To(Equal(1.0))
+		Expect(r[1]).To(Equal(1.0))
 	})
 
 	Describe("Builder", func() {
@@ -125,7 +125,7 @@ var _ = Describe("Network", func() {
 				activation.StepFunction,
 			).WithLayer(
 				[]network.Neuron{{
-					Weights: []int{0},
+					Weights: []float64{0.0},
 				}},
 			).Build()
 
@@ -139,17 +139,17 @@ var _ = Describe("Network", func() {
 				activation.StepFunction,
 			).WithLayer(
 				[]network.Neuron{{
-					Weights: []int{0, 0},
+					Weights: []float64{0.0, 0.0},
 				}, {
-					Weights: []int{0, 0},
+					Weights: []float64{0.0, 0.0},
 				}, {
-					Weights: []int{0, 0},
+					Weights: []float64{0.0, 0.0},
 				}},
 			).WithLayer(
 				[]network.Neuron{{
-					Weights: []int{0, 0, 0, 0},
+					Weights: []float64{0.0, 0.0, 0.0, 0.0},
 				}, {
-					Weights: []int{0, 0, 0, 0},
+					Weights: []float64{0.0, 0.0, 0.0, 0.0},
 				}},
 			).Build()
